+++ packages/app/src/@generic/constant/micro-unit-decimal-places.constant.ts
@@ -1,0 +1,3 @@
+import { PRECISION } from '@budgie/contracts';
+
+export const MICRO_UNIT_DECIMAL_PLACES = Math.log10(PRECISION);