+++ packages/app/src/account/interface/home-account-balance.interface.ts
@@ -1,0 +1,13 @@
+import { AccountDebtTypeEnum, AccountTypeEnum, ExternalSourceEnum } from '@budgie/contracts';
+
+export interface HomeAccountBalanceInterface {
+    readonly accountId: number;
+    readonly accountType: AccountTypeEnum;
+    readonly balance: number;
+    readonly bankProvider: ExternalSourceEnum | null;
+    readonly convertedBalance: number;
+    readonly convertedTargetBalance: number;
+    readonly debtType: AccountDebtTypeEnum;
+    readonly includeInNetWorth: boolean;
+    readonly isActive: boolean;
+}