+++ packages/app/src/account/query/use-cached-balance.query.ts
@@ -7,7 +7,8 @@ import { useCachedMicroUnitQuery } from './use-cached-micro-unit.query';
 
 type BalanceQuery =
     | ReturnType<typeof accountBalanceRepository.getByAccountId>
-    | ReturnType<typeof accountBalanceRepository.getArchivedAccountBalance>;
+    | ReturnType<typeof accountBalanceRepository.getArchivedAccountBalance>
+    | ReturnType<typeof accountBalanceRepository.getTotalByCryptoInstrument>;
 
 export const useCachedBalanceQuery = (query: BalanceQuery, dependencies: unknown[]) => {
     const accountBalancesUpdatedAt = useAccountBalancesUpdatedAtQuery();