+++ packages/contracts/src/instrument-daily-market-price/entity/instrument-daily-market-price-create-entity.interface.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod';
+
+import { InstrumentDailyMarketPriceCreateEntitySchema } from '../schema/instrument-daily-market-price-create-entity.schema';
+
+export type InstrumentDailyMarketPriceCreateEntityInterface = z.infer<typeof InstrumentDailyMarketPriceCreateEntitySchema>;