+++ packages/contracts/src/instrument-daily-market-price/entity/instrument-daily-market-price-entity.interface.ts
@@ -1,0 +1,3 @@
+import { InstrumentDailyMarketPriceEntityTable } from '../table/instrument-daily-market-price-entity.table';
+
+export type InstrumentDailyMarketPriceEntityInterface = typeof InstrumentDailyMarketPriceEntityTable.$inferSelect;